ops/scripts/requirements.txt
openai>=1.51.0
python-dotenv>=1.0.1
pypdf>=4.3.1
numpy>=1.26.0
tiktoken>=0.7.0
tenacity>=8.2.3
